Allow configuration/command-line paths other than root to be absolute
Previously FPGS would try to append all other configuration paths to
the root path, meaning that only the root path could be absolute,
while the rest were forced to be relative.

This change causes FPGS to clean, but not join, paths that are already
absolute, so that absolute paths may be used in its configuration and
on the command-line in addition to relative paths.

Path manipulation is also simplified, as multiple redundant uses of
`filepath.Abs()` + `path.Join()` were removed and replaced with
`filepath.Join()` which handles standardizing path separators and
otherwise cleaning paths to the same effect as the old method.